added chunk_web_data to chroma_vector_store.py
