+++ innovation_pathfinder_ai/vector_store/chroma_vector_store.py
@@ -12,6 +12,9 @@ import chromadb
 from langchain.text_splitter import CharacterTextSplitter
 from langchain_text_splitters import MarkdownHeaderTextSplitter
 from langchain_text_splitters import RecursiveCharacterTextSplitter
+from langchain_text_splitters.sentence_transformers import SentenceTransformersTokenTextSplitter
+
+from langchain_community.document_loaders import WebBaseLoader
 
 from langchain_core.documents import Document
 from langchain_community.document_loaders import PyPDFLoader
@@ -23,6 +26,10 @@ from langchain_community.embeddings.sentence_transformer import (
 from innovation_pathfinder_ai.utils.utils import (
     generate_uuid    
 )
+
+from typing import List, Optional
+from langchain_core.documents import Document # for typing 
+
 import dotenv
 import os
 
@@ -193,7 +200,44 @@ def add_pdf_to_vector_store(
             embeddings=embedding_function(documents_page_content[i]),
             metadatas=data.metadata,  # type: ignore
         )
+
+def chunk_web_data(
+    urls: List[str],
+    chunk_overlap: Optional[int] = 50,
+    tokens_per_chunk: Optional[int] = None,
+    model_name: str = os.getenv("EMBEDDING_MODEL"),
+    chunk_size: int = 1000,
+    ) -> List[Document]:
+    """
+    ## Summary
+    This function is used to chunk webpages
     
+    ## Arguments
+    urls list[str] : a list of urls  to be chunks
+    chunk_size int : the chunking size
+    model_name str : the embedding model used to chunk will use the environment default unless overwritten
+    tokens_per_chunk int | None : the amount of chunks per token paramter inhereted from `SentenceTransformersTokenTextSplitter`
+    chunk_size int : the size of chunks per `Document`
+
+    ## Return
+    it may be a List[Document] or None if it is a List[Document] then these chunks will be 
+    embedded wiht a different function or method
+    """
+    
+    text_splitter = SentenceTransformersTokenTextSplitter(
+        chunk_overlap=chunk_overlap,
+        tokens_per_chunk=tokens_per_chunk,
+        model_name=model_name,
+        chunk_size=chunk_size
+    )
+    
+    loader = WebBaseLoader(urls)
+
+    data = loader.load_and_split(
+        text_splitter=text_splitter
+    )
+    
+    return data
     
 if __name__ == "__main__":
